% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-fn.R
\name{get_dt_from_env}
\alias{get_dt_from_env}
\title{Get dataframe/datatable list from calling environment}
\usage{
get_dt_from_env()
}
\value{
char vector of dt list
}
\description{
This returns a list of all dataframes / datatables in the calling environment.
}
\details{
(Internal function)
}
