% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-fn.R
\name{process_dt_from_env}
\alias{process_dt_from_env}
\title{Process dataframe selected from environment}
\usage{
process_dt_from_env(out)
}
\arguments{
\item{out}{raw dataframe from env}
}
\value{
list of outputs
}
\description{
This internal function is responsible for pre-processing a selected df from the env. The shiny app
relies on this func to check all req of the df (cols, col types, data quality etc), and req that 
this function gives an error on violating any requirements.
}
\details{
(Internal function)
}
