% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-fn.R
\name{process_input_file}
\alias{process_input_file}
\title{Processes input CSV/TSV files}
\usage{
process_input_file(input_file, sep, quote, ANOMALY_TAGS, GROUPS, date_coltype)
}
\arguments{
\item{input_file}{}

\item{sep}{}

\item{quote}{}

\item{ANOMALY_TAGS}{}

\item{GROUPS}{}

\item{date_coltype}{}
}
\value{

}
\description{
This file would contain the timeseries to be labeled. A typical file will contain a minimum of 3 columns
(ds, grp, value)
}
\details{
The ds column may be date (%Y-%m-%d) or date-time (%Y-%m-%d %H:%M:%S). `UTC` timezone is assumed
}
